# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Cycle struct {
	ID               int64              `json:"id"`
	UserID           pgtype.UUID        `json:"user_id"`
	StartTime        pgtype.Timestamptz `json:"start_time"`
	EndTime          pgtype.Timestamptz `json:"end_time"`
	TimezoneOffset   pgtype.Interval    `json:"timezone_offset"`
	Strain           pgtype.Float4      `json:"strain"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
	Kilojoule        pgtype.Float4      `json:"kilojoule"`
	AverageHeartRate pgtype.Int4        `json:"average_heart_rate"`
	MaxHeartRate     pgtype.Int4        `json:"max_heart_rate"`
}

type DailyRecovery struct {
	UserID      pgtype.UUID `json:"user_id"`
	Bucket      interface{} `json:"bucket"`
	AvgRecovery float64     `json:"avg_recovery"`
}

type DailyStrain struct {
	UserID    pgtype.UUID `json:"user_id"`
	Bucket    interface{} `json:"bucket"`
	AvgStrain float64     `json:"avg_strain"`
	MaxStrain interface{} `json:"max_strain"`
}

type Recovery struct {
	ID               int64              `json:"id"`
	UserID           pgtype.UUID        `json:"user_id"`
	StartTime        pgtype.Timestamptz `json:"start_time"`
	TimezoneOffset   pgtype.Interval    `json:"timezone_offset"`
	RecoveryScore    pgtype.Float4      `json:"recovery_score"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
	RestingHeartRate pgtype.Float4      `json:"resting_heart_rate"`
	HrvRmssdMilli    pgtype.Float4      `json:"hrv_rmssd_milli"`
	Spo2Percentage   pgtype.Float4      `json:"spo2_percentage"`
	SkinTempCelsius  pgtype.Float4      `json:"skin_temp_celsius"`
}

type Sleep struct {
	ID                          int64              `json:"id"`
	UserID                      pgtype.UUID        `json:"user_id"`
	StartTime                   pgtype.Timestamptz `json:"start_time"`
	EndTime                     pgtype.Timestamptz `json:"end_time"`
	TimezoneOffset              pgtype.Interval    `json:"timezone_offset"`
	PerformanceScore            pgtype.Float4      `json:"performance_score"`
	CreatedAt                   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt                   pgtype.Timestamptz `json:"updated_at"`
	Nap                         pgtype.Bool        `json:"nap"`
	RespiratoryRate             pgtype.Float4      `json:"respiratory_rate"`
	SleepConsistencyPercentage  pgtype.Float4      `json:"sleep_consistency_percentage"`
	SleepEfficiencyPercentage   pgtype.Float4      `json:"sleep_efficiency_percentage"`
	SleepDebtMilli              pgtype.Int4        `json:"sleep_debt_milli"`
	TotalInBedTimeMilli         pgtype.Int4        `json:"total_in_bed_time_milli"`
	TotalAwakeTimeMilli         pgtype.Int4        `json:"total_awake_time_milli"`
	TotalNoDataTimeMilli        pgtype.Int4        `json:"total_no_data_time_milli"`
	TotalLightSleepTimeMilli    pgtype.Int4        `json:"total_light_sleep_time_milli"`
	TotalSlowWaveSleepTimeMilli pgtype.Int4        `json:"total_slow_wave_sleep_time_milli"`
	TotalRemSleepTimeMilli      pgtype.Int4        `json:"total_rem_sleep_time_milli"`
	SleepCycleCount             pgtype.Int4        `json:"sleep_cycle_count"`
	DisturbanceCount            pgtype.Int4        `json:"disturbance_count"`
}

type User struct {
	ID                    pgtype.UUID        `json:"id"`
	WhoopUserID           string             `json:"whoop_user_id"`
	EncryptedAccessToken  []byte             `json:"encrypted_access_token"`
	EncryptedRefreshToken []byte             `json:"encrypted_refresh_token"`
	CreatedAt             pgtype.Timestamptz `json:"created_at"`
	UpdatedAt             pgtype.Timestamptz `json:"updated_at"`
}

type WebhookEvent struct {
	ID        pgtype.UUID        `json:"id"`
	Payload   []byte             `json:"payload"`
	Status    string             `json:"status"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type WeeklyRecovery struct {
	UserID      pgtype.UUID `json:"user_id"`
	Bucket      interface{} `json:"bucket"`
	AvgRecovery float64     `json:"avg_recovery"`
}

type WeeklyStrain struct {
	UserID    pgtype.UUID `json:"user_id"`
	Bucket    interface{} `json:"bucket"`
	AvgStrain float64     `json:"avg_strain"`
	MaxStrain interface{} `json:"max_strain"`
}

type Workout struct {
	ID                  int64              `json:"id"`
	UserID              pgtype.UUID        `json:"user_id"`
	StartTime           pgtype.Timestamptz `json:"start_time"`
	EndTime             pgtype.Timestamptz `json:"end_time"`
	TimezoneOffset      pgtype.Interval    `json:"timezone_offset"`
	SportID             pgtype.Int4        `json:"sport_id"`
	Strain              pgtype.Float4      `json:"strain"`
	CreatedAt           pgtype.Timestamptz `json:"created_at"`
	UpdatedAt           pgtype.Timestamptz `json:"updated_at"`
	AverageHeartRate    pgtype.Int4        `json:"average_heart_rate"`
	MaxHeartRate        pgtype.Int4        `json:"max_heart_rate"`
	Kilojoule           pgtype.Float4      `json:"kilojoule"`
	PercentRecorded     pgtype.Float4      `json:"percent_recorded"`
	DistanceMeter       pgtype.Float4      `json:"distance_meter"`
	AltitudeGainMeter   pgtype.Float4      `json:"altitude_gain_meter"`
	AltitudeChangeMeter pgtype.Float4      `json:"altitude_change_meter"`
	ZoneZeroMilli       pgtype.Int4        `json:"zone_zero_milli"`
	ZoneOneMilli        pgtype.Int4        `json:"zone_one_milli"`
	ZoneTwoMilli        pgtype.Int4        `json:"zone_two_milli"`
	ZoneThreeMilli      pgtype.Int4        `json:"zone_three_milli"`
	ZoneFourMilli       pgtype.Int4        `json:"zone_four_milli"`
	ZoneFiveMilli       pgtype.Int4        `json:"zone_five_milli"`
}